requirements.txt
Flask>=2.0.0
paho-mqtt>=1.6.1
Flask-SocketIO>=5.0.0
python-dotenv>=0.19.0
requests>=2.25.0
numpy>=1.20.0
pandas>=1.3.0
python-engineio>=4.0.0
python-socketio>=5.0.0